+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Option to mute chat reactions.
+- Notify reactions only on local accounts messages.
 
 ## 5.2.3 - 2024-03-29